back/requirements.txt
torch
datasets
pyyaml
numpy<2